refactor: move more and more components to TS/styled components; add MobileSidebar and use in all about and events pages
