+++ src/pages/about/history.js
@@ -9,6 +9,7 @@ import Container from '../../components/shared/container';
 import FuturaParagraph from '../../components/shared/futura-paragraph';
 import Layout from '../../components/layout';
 import presets from '../../utils/presets';
+import MobileSidebar from '../../components/shared/MobileSidebar';
 import SidebarBody from '../../components/shared/sidebar/SidebarBody';
 
 // Sidebar data
@@ -75,27 +76,12 @@ const History = ({ location }) => (
         <div>Director of Fine Arts, Retired</div>
         <div>Irving ISD, Irving, Texas</div>
 
-        {/* Mobile sidebar */}
-        <div
-          css={{
-            display: 'block',
-            [presets.Tablet]: {
-              display: 'none',
-            },
-          }}
-        >
-          <hr
-            css={{
-              border: 0,
-              height: 2,
-              marginTop: 10,
-            }}
-          />
+        <MobileSidebar>
           <SidebarBody
             inline
             yaml={aboutSidebar}
           />
-        </div>
+        </MobileSidebar>
       </Container>
     </div>
   </Layout>
