+++ src/pages/about/philosophy.js
@@ -9,6 +9,7 @@ import Cards from '../../components/shared/cards';
 import Container from '../../components/shared/container';
 import FuturaParagraph from '../../components/shared/futura-paragraph';
 import Layout from '../../components/layout';
+import MobileSidebar from '../../components/shared/MobileSidebar';
 import SidebarBody from '../../components/shared/sidebar/SidebarBody';
 
 // Sidebar data
@@ -175,27 +176,13 @@ const Philosophy = ({ location }) => (
             </FuturaParagraph>
           </PhilosophyCard>
         </Cards>
-        {/* Mobile sidebar */}
-        <div
-          css={{
-            display: 'block',
-            [presets.Tablet]: {
-              display: 'none',
-            },
-          }}
-        >
-          <hr
-            css={{
-              border: 0,
-              height: 2,
-              marginTop: 10,
-            }}
-          />
+
+        <MobileSidebar>
           <SidebarBody
             inline
             yaml={aboutSidebar}
           />
-        </div>
+        </MobileSidebar>
       </Container>
     </div>
   </Layout>
