+++ src/pages/events/tmea-round-table.js
@@ -6,8 +6,8 @@ import React from 'react';
 // Internal Dependencies
 import Container from '../../components/shared/container';
 import Layout from '../../components/layout';
+import MobileSidebar from '../../components/shared/MobileSidebar';
 import SidebarBody from '../../components/shared/sidebar/SidebarBody';
-import presets from '../../utils/presets';
 import { useEventData } from '../../utils/hooks/useEventData';
 
 // Sidebar data
@@ -71,26 +71,13 @@ const TmeaRoundTable = ({ location }) => {
             convention.
           </p>
         </section>
-        <div
-          css={{
-            display: 'block',
-            [presets.Tablet]: {
-              display: 'none',
-            },
-          }}
-        >
-          <hr
-            css={{
-              border: 0,
-              height: 2,
-              marginTop: 10,
-            }}
-          />
+
+        <MobileSidebar>
           <SidebarBody
             inline
             yaml={eventsSidebar}
           />
-        </div>
+        </MobileSidebar>
       </Container>
     </Layout>
   );
